double

+++ src/workbook/StepA/Apartment.java
@@ -14,7 +14,7 @@ public class Apartment {
 	void input() {
 		Scanner s = new Scanner(System.in);
 		System.out.printf("아파트의 분양 면적을 입력하시오.");
-		this.m2_area = s.nextInt();
+		this.m2_area = s.nextDouble();
 	}	
 	double getArea() {
 		pyung_area = m2_area / 3.305;
